intial commit for the gate library specific utils plugin and implementing it into dana.
