+++ plugins/dataflow_analysis/CMakeLists.txt
@@ -19,7 +19,7 @@ if(PL_DATAFLOW OR BUILD_ALL_PLUGINS)
                    SOURCES ${DATAFLOW_SRC} ${DATAFLOW_PYTHON_SRC}
                    INCLUDES PUBLIC $<BUILD_INTERFACE:${IGRAPH_INCLUDES}>
                    DEFINITIONS PUBLIC -DIGRAPH_VERSION_MAJOR_GUESS=${IGRAPH_VERSION_MAJOR_GUESS} -DIGRAPH_VERSION_MINOR_GUESS=${IGRAPH_VERSION_MINOR_GUESS} -DIGRAPH_VERSION_PATCH_GUESS=${IGRAPH_VERSION_PATCH_GUESS}
-                   LINK_LIBRARIES PUBLIC ${IGRAPH_LIBRARIES} graph_algorithm
+                   LINK_LIBRARIES PUBLIC ${IGRAPH_LIBRARIES} graph_algorithm gate_library_specific_utils
                    )
 
 endif()
