+++ plugins/dataflow_analysis/include/dataflow_analysis/plugin_dataflow.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "dataflow_analysis/processing/configuration.h"
-#include "dataflow_analysis/utils/utils.h"
 #include "hal_core/plugin_system/plugin_interface_cli.h"
 
 #include <vector>
