+++ plugins/gate_library_specific_utils/CMakeLists.txt
@@ -1,0 +1,12 @@
+option(PL_GATE_LIBRARY_SPECIFIC_UTILS "PL_GATE_LIBRARY_SPECIFIC_UTILS" ON)
+if(PL_GATE_LIBRARY_SPECIFIC_UTILS OR BUILD_ALL_PLUGINS)
+    file(GLOB_RECURSE GATE_LIBRARY_SPECIFIC_UTILS_INC ${CMAKE_CURRENT_SOURCE_DIR}/include/*.h)
+    file(GLOB_RECURSE GATE_LIBRARY_SPECIFIC_UTILS_SRC ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
+
+    hal_add_plugin(gate_library_specific_utils
+                   SHARED
+                   HEADER ${GATE_LIBRARY_SPECIFIC_UTILS_INC}
+                   SOURCES ${GATE_LIBRARY_SPECIFIC_UTILS_SRC}
+                   LINK_LIBRARIES PUBLIC
+                   )
+endif()
